Adding my Gemini server Titan
Hello!
I'm submitting my Gemini server Titan; it's still under construction but it's in a working state now and I'd love input on how to improve it from the community! If it can help anyone out there that would be pretty cool too :) Idk if it's awesome but I'm pretty proud of what I've done so far. Thanks!