Include link to the Gempub specification
Include a link to the Gempub specification at https://codeberg.org/oppenlab/gempub. Gempub is an e-book and archive container specification. It is intended to serve as a container for e-books containing text files of in the "text/gemini" format, allowing the author to avoid the complexity of using web technologies for e-books that do not require them. It has a secondary purpose of functioning as an archival container for Gemini capsules.